refactor: move analytics controller under internal path
