+++ src/main/kotlin/com/example/jaebitly/application/RedirectStatStore.kt
@@ -3,6 +3,10 @@ package com.example.jaebitly.application
 import com.example.jaebitly.domain.ShortKey
 import com.example.jaebitly.domain.stat.RedirectStat
 
+/**
+ * Internal store for redirect analytics.
+ * Implementation may change or be removed.
+ */
 interface RedirectStatStore {
     fun save(stat: RedirectStat)
